massive update
basically a full restructure of the site. Moving over to collections, as those are much easier when it comes to adding new items and pages.

other:
removed a bunch of unnecessary files
simpilified the bootleg covers so that only one artist file is needed for each
moved a few project pages to blog posts instead, might move those back to collections not sure